+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * add data-remove processing directive
 * add data-clear processing directive
+* add *-closest variants of data-remove and data-clear
 * remove processError and replaced with triggering bootstrap-ajax:error with
   the element and statusCode as parameters